Transferred from the Adobe Open Source Subversion Repository
Revision: 7766
Author:   dloverin@adobe.com
Date:     2009-06-11T23:14:47.510089Z

Log Message: 
-----------
WindowedApplication and Window bug fixes.

QE Notes: None.
Doc Notes: Changed some ASDoc comments.
Bugs: SDK-21672, SDK-21625, SDK-21628
Reviewer: Glenn
tests: checkintests

On 6/11/09 6:54 AM, "Darrell Loverin" <dloverin@adobe.com> wrote:
http://bugs.adobe.com/jira/browse/SDK-21672 When using system chrome, setting height to 0 moves the status bar to the top of the app. and does not resize the app.
 
Modified commitProperties(). This looks like an AIR bug. I worked around the problem by first setting the nativeWindow to the correct size and then setting the stage.stageHeight.

http://bugs.adobe.com/jira/browse/SDK-21625 minWidth and minHeight default to incorrect values.
Modified ASDoc @default comment for minWidth, minHeight, maxWidth, and maxHeight.
 
http://bugs.adobe.com/jira/browse/SDK-21628 minWidth and minHeight override the width and height when set in MXML.
 
The height and width are set by the auto generated code before the application window is created. The problem was that maxHeight is initialized to zero and is used if there is no nativeWindow. So when the height property is set to 500 and the autogenerated code sets it before the nativeWindow is created, the height setting will see the height is greater than maxHeight and set it to zero. 
 
I fixed the problem by initializing _maxHeight and _maxWidth to 2880 (the documented maximum width and height of an AIR window).

Modified Paths:
--------------
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/airframework/src/mx/core/Window.as</path>
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/airframework/src/spark/components/WindowedApplication.as</path>
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/airframework/src/spark/components/Window.as</path>
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/airframework/src/mx/core/WindowedApplication.as</path>

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1249849 13f79535-47bb-0310-9956-ffa450edef68