Transferred from the Adobe Open Source Subversion Repository
Revision: 10178
Author:   rfrishbe@adobe.com
Date:     2009-09-11T22:36:24.712602Z

Log Message: 
-----------
This is a bug fix for setting AIR Window/WindowedApplication visibility when a hideEffect is installed.  We need to dispatch the HIDE event for the effect to play.  Then, when the effect is finished, we can remove that event listener and set the visibility to false.

I've already checked the fixes in for the Halo versions in to 3.x with 10177.

QE notes: -
Doc notes: -
Bugs: SDK-14615
Reviewer: Darrell
Tests run: checkintests, Rob helped run the mustella tests for Window/WindowedApplication
Is noteworthy for integration: no

Modified Paths:
--------------
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/airframework/src/spark/components/WindowedApplication.as</path>
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/airframework/src/spark/components/Window.as</path>

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1250525 13f79535-47bb-0310-9956-ffa450edef68