Transferred from the Adobe Open Source Subversion Repository
Revision: 8192
Author:   dloverin@adobe.com
Date:     2009-06-24T19:37:55.466313Z

Log Message: 
-----------
Fix resizing a window with the gripper.

WindowedApplication.as, Window.as
After the click on the gripper is detected and the resized is started just return. The code was falling into the hit testing for the application border.

QE Notes: none.
Doc Notes: none.
Bugs: SDK-21631
Reviewer: Glenn
tests: checkintests

Modified Paths:
--------------
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/airframework/src/spark/components/WindowedApplication.as</path>
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/airframework/src/spark/components/Window.as</path>

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1250573 13f79535-47bb-0310-9956-ffa450edef68