Transferred from the Adobe Open Source Subversion Repository
Revision: 11789
Author:   rfrishbe@adobe.com
Date:     2009-11-13T19:38:40.358315Z

Log Message: 
-----------
When a UIMovieClip resizes, it does so through scaling. Before UIMC was keeping track of this on it's own.  Now that AdvancedLayoutFeatures support this through scaleX/scaleY, UIMC can now rely on this behavior.  This also helps fix some Effects bugs which rely on this to figure out the difference in transformations that get applied and contribute to the "width" and "height" properties and ones that don't.

QE notes: Can re-include the 3 FCK tests and regen the bitmaps
Doc notes: -
Bugs: SDK-23754, SDK-23779
Reviewer: Evtim
Tests run: checkintest, mustella FCK
Is noteworthy for integration: No

Modified Paths:
--------------
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/flash-integration/src/mx/flash/UIMovieClip.as</path>
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/flash-integration/src/mx/flash/FlexContentHolder.as</path>
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/utilities/src/MyFlexContentHolder.as</path>
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/flash-integration/src/mx/flash/ContainerMovieClip.as</path>

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1253693 13f79535-47bb-0310-9956-ffa450edef68