Transferred from the Adobe Open Source Subversion Repository
Revision: 13318
Author:   rfrishbe@adobe.com
Date:     2010-01-07T01:36:07.764515Z

Log Message: 
-----------
Fixing a Flash Component Kit bug where the bounds of a container was larger than it should be when the container was sized smaller than it's natural size and scaleContentWhenResized=false.  This is because the "Flex Content Symbol" was getting resized and scaled out of the bounds.  To counter-act this, we now scale the symbol when appropriate.

I'm also updated the MXP file for the Flash Component Kit.

QE notes: -
Doc notes: -
Bugs: SDK-24930
Reviewer: Glenn, Chet
Tests run: checkintests, mustella FCK
Is noteworthy for integration: No

Modified Paths:
--------------
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/flash-integration/src/mx/flash/FlexContentHolder.as</path>
<path kind="file" action="M">/flex/sdk/trunk/frameworks/flash-integration/FlexComponentKit.mxp</path>

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1253711 13f79535-47bb-0310-9956-ffa450edef68