Transferred from the Adobe Open Source Subversion Repository
Revision: 10961
Author:   chaase@adobe.com
Date:     2009-10-09T21:55:11.382317Z

Log Message: 
-----------
Fix for occasional RTE in transform effects. The problem was that an object whose parent container was 3D transformed which then underwent a 2D transform effect might try to access the 2D matrix of the display object. In some situations, that 2D matrix is null, because the 3D matrix on the display object is the one currently being used. 
The fix is to allocate AdvancedLayoutFeatures and let that class handle this matrix abstraction for us.

QE notes: None
Doc notes: No
Bugs: sdk-23546
Reviewer: Evtim
Tests run: checkintests, mustella spark/effects
Is noteworthy for integration: Yes

Modified Paths:
--------------
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/spark/src/spark/core/SpriteVisualElement.as</path>
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/flash-integration/src/mx/flash/UIMovieClip.as</path>
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/framework/src/mx/core/UIComponent.as</path>

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1253754 13f79535-47bb-0310-9956-ffa450edef68