Transferred from the Adobe Open Source Subversion Repository
Revision: 7573
Author:   gosmith@adobe.com
Date:     2009-06-05T06:20:06.217095Z

Log Message: 
-----------
More changes to text styles.

The global selector for Flex 4 now sets the default value for 'kerning' to "off" instead of false. The legal values in Halo components are true and false. The legal values in Spark components are "on", "off", and "auto". In order to support any of these five values in a selector that affects both Halo and Spark components, Halo components map "on" and "auto" to true and "off" to false, and Spark components map true to "on" and false to "off".

We will decide later whether to switch the default to "auto", which is FTE and TLF's default.

The global selector for Flex 4 now sets the default value for 'textAlign' to "start" instead of "left". The legal values in Halo components are "left", "center", "right", and "justify", but Spark components add "start" and "end" to support RTL text. In order to support any of these six values in a selector that affects both Halo and Spark components, Halo components map "start" to "left" and "end" to "right".

QE Notes: Some tests may need updating due to these changes.
Doc Notes: None
Bugs: SDK-21624
Reviewer: Carol

Modified Paths:
--------------
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/framework/src/mx/controls/Button.as</path>
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/framework/src/mx/core/UITextField.as</path>
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/framework/defaults.css</path>

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1253890 13f79535-47bb-0310-9956-ffa450edef68