Transferred from the Adobe Open Source Subversion Repository
Revision: 14999
Author:   klin@adobe.com
Date:     2010-03-24T20:33:13.171795Z

Log Message: 
-----------
Merge revision 14996 from 4.x to trunk.

Update for implementing ILayoutDirectionElement in BitmapAsset and SpriteAsset. Instead of trying to handle transforming the assets directly, we leverage AdvancedLayoutFeatures to update the transform matrix of the assets. The main difference between the asset implementations and other ones is that width/height and scaleX/Y are tied to each other. The setters of each of those properties will set the backing value of their respective property. Another change is that we use the existence of layoutFeatures to keep track of whether the asset needs to de-mirror instead of using the mirror variable.

Also, I removed a FIXME from AdvancedLayoutFeatures. We already short-circuit this in the classes that own the AdvancedLayoutFeatures. I changed the FIXMEs for the x getters to TODOs.

QE notes: No
Doc notes: No
Bugs: SDK-25879, SDK-25894
Reviewer: Hans, Carol, Evtim
Tests run: checkintests
Is noteworthy for integration: No

Modified Paths:
--------------
<path kind="dir" action="M">/flex/sdk/trunk</path>
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/framework/src/mx/core/SpriteAsset.as</path>
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/framework/src/mx/core/UITextField.as</path>
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/spark/src/mx/core/UIFTETextField.as</path>
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/framework/src/mx/core/BitmapAsset.as</path>
<path kind="file" action="M">/flex/sdk/trunk/frameworks/projects/framework/src/mx/core/AdvancedLayoutFeatures.as</path>

git-svn-id: https://svn.apache.org/repos/asf/incubator/flex/trunk/frameworks@1259004 13f79535-47bb-0310-9956-ffa450edef68